renamed urls and fetches files, made props readonly
