+++ src/components/EditPasswordDialog.tsx
@@ -4,7 +4,7 @@ import { SubmitHandler, useForm } from "react-hook-form";
 import { useMutation } from "react-query";
 import * as yup from "yup";
 import { fetchPut } from "../utils/Fetches";
-import { urls } from "../utils/urls";
+import { urls } from "../utils/Urls";
 import { useContext } from "react";
 import { UserContext } from "../contexts/UserContext";
 import { QuizContext } from "../contexts/QuizContext";
@@ -19,8 +19,8 @@ const schema = yup.object({
 type Inputs = yup.InferType<typeof schema>;
 
 interface EditPasswordDialogProps {
-    open: boolean
-    close: () => void
+    readonly open: boolean
+    readonly close: () => void
 }
 
 export default function EditPasswordDialog(props: EditPasswordDialogProps) {
